Move hasIntersects into its own file
